requirements.txt
pandas==1.4.1
biopython==1.79
numpy==1.22.3
gensim==4.1.2
torch==1.10.1
torchvision==0.11.2
torchaudio==0.10.1
joblib==1.1.0
scikit-learn==1.0.2
